refactor: unnested Explorer View components
