+++ app/components/Explorer/ExplorerBody/ExplorerCodeView.js
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames';
 import ace from 'brace';
 import { connect } from "react-redux";
-import { saveRow } from '../../../../actions';
+import { saveRow } from '../../../actions';
 
 // Looks weird, but this is necessary to apply the theme to ace editor
 require('brace/mode/json');
