+++ app/components/Explorer/ExplorerBody/ExplorerTableView/ExplorerTableView.js
@@ -6,10 +6,9 @@ import _ from 'lodash';
 import classNames from 'classnames';
 import { connect } from 'react-redux';
 import ExplorerTableCell from './ExplorerTableCell.js';
-import { refreshExplorerBody } from '../../../../../actions';
 import { getColumnNames, getColumnWidth } from './explorer-table-view-utils';
-import { saveRow, saveInlineEdit } from '../../../../../actions';
-import * as types from '../../../../../action-types';
+import { saveRow, saveInlineEdit, refreshExplorerBody } from '../../../../actions';
+import * as types from '../../../../action-types';
 
 class ExplorerTableView extends Component {
